somc-wifi-ctrl: return all 0xFF as the default MAC
The first call to this function is before the MAC is set and will be
used as the "permanent address". NetworkManager uses this to distinguish
between interfaces, unless it's "invalid" like all 0x00 or 0xFF. Putting
a random MAC here will confuse NM, thus put all 0xFF instead so that NM
will failback to normal interface MAC address.